refactor(upload): component optimization
