+++ packages/components/upload/components/multiple-flow-list.tsx
@@ -16,7 +16,7 @@ import { isFunction } from 'lodash-es';
 import { isObject } from 'lodash-es';
 import useGlobalIcon from '../../hooks/useGlobalIcon';
 import ImageViewer, { ImageViewerProps } from '../../image-viewer';
-import { CommonDisplayFileProps } from '../interface';
+import { CommonDisplayFileProps } from '../types';
 import { commonProps } from '../constants';
 import TButton from '../../button';
 import { UploadFile, TdUploadProps } from '../type';
