+++ packages/components/upload/index.ts
@@ -3,7 +3,7 @@ import withInstall from '../utils/withInstall';
 
 import './style';
 
-export * from './interface';
+export * from './types';
 
 export const Upload = withInstall(_Upload);
 export default Upload;
